Organized and re-structured frontend folders and files
