+++ frontend/src/App.jsx
@@ -6,24 +6,24 @@
 * - Makes auth functions available app-wide via useAuth hook
 */
 
-import { AuthProvider } from "./components/AuthContext.jsx";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Homepage from "./components/Homepage.jsx";
-import Signin from "./components/Signin.jsx";  
-import Signup from "./components/Signup.jsx"; 
-import Profile from "./components/profile/Profile.jsx";
-import ExercisePage from "./components/ExercisePage.jsx";
+import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
 import "./App.css";
+import { AuthProvider } from "./components/auth/AuthContext.jsx";
+import ExercisePage from "./components/pages/ExercisePage.jsx";
+import HomePage from "./components/pages/Homepage.jsx";
+import ProfilePage from "./components/pages/ProfilePage.jsx";
+import Signin from "./components/pages/Signin.jsx";
+import Signup from "./components/pages/Signup.jsx";
 
 export const App = () => {
   return (
     <AuthProvider>
       <Router>
         <Routes>
-          <Route path="/" element={<Homepage />} />
+          <Route path="/" element={<HomePage />} />
           <Route path="/auth/signin" element={<Signin />} /> {/* Sign In */}
           <Route path="/auth/signup" element={<Signup />} /> {/* Sign Up */}
-          <Route path="/profile" element={<Profile />} /> {/* Profile Page */}
+          <Route path="/profile" element={<ProfilePage />} /> {/* Profile Page */}
           <Route path="/exercises" element={<ExercisePage />} />
         </Routes>
       </Router>
