+++ frontend/src/components/pages/ExercisePage.css
